+++ spec/spec_helper.rb
@@ -7,6 +7,5 @@ require 'rspec-given'
 SimpleCov.start
 
 require 'event_bus'
-require 'event_bus/testing'
